fixed scaling for gallery
